chore: updateislands
